+++ terrascript/__init__.py
@@ -74,7 +74,7 @@ class Terrascript(object):
 
         return self
 
-    
+
     def add(self, item):
         self.__add__(item)
         return item
@@ -86,7 +86,9 @@ class Terrascript(object):
 
         def _json_default(v):
             # How to encode non-standard objects
-            if isinstance(v, UserDict):
+            if isinstance(v, provisioner):
+                return {v._type: v.data}
+            elif isinstance(v, UserDict):
                 return v.data
             else:
                 return str(v)
@@ -119,7 +121,7 @@ class Terrascript(object):
         proc = subprocess.Popen(['terraform','validate','-check-variables=false'], cwd=tmpdir,
                                 stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         proc.communicate()
-        
+
         tmpfile.close()
 
         return proc.returncode == 0
@@ -171,12 +173,7 @@ class _base(object):
     @property
     def interpolated(self):
         """The object in interpolated syntax: ``${...}``."""
-        if self._class == 'variable':
-            return '${{{}}}'.format(self.fullname)
-        elif self._class == 'resources':
-            return '${{{}}'.format(self._fullname)
-        else:
-            return '${{{}}'.format(self._fullname)
+        return '${{{}}}'.format(self.fullname)
 
     @property
     def fullname(self):
@@ -198,14 +195,8 @@ class _data(_base):
     """Base class for data sources."""
     _class = 'data'
 
-    # # TODO: Work-around for https://github.com/mjuenema/python-terrascript/issues/3
-    # def __init__(self, name, **kwargs):
-    #     if kwargs:
-    #         if not 'type' in kwargs:
-    #             kwargs['type'] = 'string'
-    #         if not 'description' in kwargs:
-    #             kwargs['description'] = ''
-    #     super(_data, self).__init__(name, **kwargs)
+    def __init__(self, obj_name, **kwargs):
+        super(_data, self).__init__(obj_name, **kwargs)
 
 
 class resource(_base):
@@ -254,8 +245,10 @@ class terraform(_base):
         super(terraform, self).__init__(None, **kwargs)
 
 
-class provisioner(_base):
-    _class = 'provisioner'
+class provisioner(UserDict):
+    def __init__(self, type_, **kwargs):
+       self._type = type_
+       self.data = kwargs
 
 
 class connection(UserDict):