+++ terrascript/vsphere/d.py
@@ -5,6 +5,24 @@ datacenter = vsphere_datacenter
 class vsphere_distributed_virtual_switch(_data): pass
 distributed_virtual_switch = vsphere_distributed_virtual_switch
 
+class vsphere_virtual_machine(_data): pass
+virtual_machine = vsphere_virtual_machine
+
+class vsphere_custom_attribute(_data): pass
+custom_attribute = vsphere_custom_attribute
+
+class vsphere_datastore(_data): pass
+datastore = vsphere_datastore
+
+class vsphere_datastore_cluster(_data): pass
+datastore_cluster = vsphere_datastore_cluster
+
+class vsphere_compute_cluster(_data): pass
+compute_cluster = vsphere_compute_cluster
+
+class vsphere_resource_pool(_data): pass
+resource_pool = vsphere_resource_pool
+
 class vsphere_host(_data): pass
 host = vsphere_host
 
@@ -19,4 +37,3 @@ tag_category = vsphere_tag_category
 
 class vsphere_vmfs_disks(_data): pass
 vmfs_disks = vsphere_vmfs_disks
-