+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## [Unreleased]
 ### Fixed
-* Work-around for data/JSON bug  (issue 3).
+* Work-around for data/JSON bug  (issue #3).
 ### Added
-* `Terrascript.add()` method (issue 33).
+* `Terrascript.add()` method (issue #33).
+* Missing vSphere data sources and resources (issue #35).
 ### Removed
-*  Dropped support for Terraform < 0.10.6 (issue 38).
+*  Dropped support for Terraform < 0.10.6 (issue #38).
 
 ## [0.5.0]
 * Added `terrascript.Terrascript` class wich makes this release